# lib/fs: Optimize TempName + some cosmetic changes